update ssh2 move to vmssoftware filter remote files by version do not repeat waitable operation emit 'continue' by self purge after synchronize

+++ src/ssh-helper/stream/sftp-client.ts
@@ -82,20 +82,24 @@ export class SftpClient extends SshClient {
         await this.waitOperation.acquire();
         let statRet: IStats | undefined;
         if (await this.ensureSftp()) {
-            if (this.sftp) {
+            if (this.client && this.sftp) {
                 const opName = localize("operation.getstat", "get stat {0} via sftp{1}", file, this.tag ? " " + this.tag : "");
-                await WaitableOperation(opName, this.sftp, "continue", this.sftp, "error", (complete) => {
+                await WaitableOperation(opName, this.client, "continue", this.sftp, "error", (complete) => {
                     if (!this.sftp) {
                         return false;
                     }
-                    return !this.sftp.stat(file, (err, stat) => {
+                    let should_wait = !this.sftp.stat(file, (err, stat) => {
                         if (err) {
                             this.logFn(LogType.debug, () => localize("debug.operation.error", "{0} error: {1}", opName, err.message));
                         } else {
                             statRet = stat;
                         }
                         complete.release();
+                        if (should_wait) {
+                            this.client?.emit("continue");
+                        }
                     });
+                    return should_wait;
                 }, this.logFn);
             }
         }
@@ -107,20 +111,24 @@ export class SftpClient extends SshClient {
         let fileDeleted = false;
         await this.waitOperation.acquire();
         if (await this.ensureSftp()) {
-            if (this.sftp) {
+            if (this.client && this.sftp) {
                 const opName = localize("operation.delete", "delete {0} via sftp{1}", file, this.tag ? " " + this.tag : "");
-                await WaitableOperation(opName, this.sftp, "continue", this.sftp, "error", (complete) => {
+                await WaitableOperation(opName, this.client, "continue", this.sftp, "error", (complete) => {
                     if (!this.sftp) {
                         return false;
                     }
-                    return !this.sftp.unlink(file, (err) => {
+                    let should_wait = !this.sftp.unlink(file, (err) => {
                         if (err) {
                             this.logFn(LogType.debug, () => localize("debug.operation.error", "{0} error: {1}", opName, err.message));
                         } else {
                             fileDeleted = true;
                         }
                         complete.release();
+                        if (should_wait) {
+                            this.client?.emit("continue");
+                        }
                     });
+                    return should_wait;
                 }, this.logFn);
             }
         }
@@ -131,18 +139,22 @@ export class SftpClient extends SshClient {
     public async setStat(file: string, stat: IInputAttributes) {
         await this.waitOperation.acquire();
         if (await this.ensureSftp()) {
-            if (this.sftp) {
+            if (this.client && this.sftp) {
                 const opName = localize("operation.setstat", "set stat {0} via sftp{1}", file, this.tag ? " " + this.tag : "");
-                await WaitableOperation(opName, this.sftp, "continue", this.sftp, "error", (complete) => {
+                await WaitableOperation(opName, this.client, "continue", this.sftp, "error", (complete) => {
                     if (!this.sftp) {
                         return false;
                     }
-                    return !this.sftp.setstat(file, stat, (err) => {
+                    let should_wait = !this.sftp.setstat(file, stat, (err) => {
                         if (err) {
                             this.logFn(LogType.debug, () => localize("debug.operation.error", "{0} error: {1}", opName, err.message));
                         }
                         complete.release();
+                        if (should_wait) {
+                            this.client?.emit("continue");
+                        }
                     });
+                    return should_wait;
                 }, this.logFn);
             }
         }
@@ -153,22 +165,26 @@ export class SftpClient extends SshClient {
     public async readDirectory(directory: string) {
         await this.waitOperation.acquire();
         let files: FileEntry[] | undefined;
-        if (await this.ensureSftp() && this.sftp) {
+        if (await this.ensureSftp() && this.client && this.sftp) {
             // set empty list => return 'undefined' only if stfp fails, but not if directory doesn't exist
             files = [];
             const opName = localize("operation.readdir", "read directory {0} via sftp{1}", directory, this.tag ? " " + this.tag : "");
-            await WaitableOperation(opName, this.sftp, "continue", this.sftp, "error", (complete) => {
+            await WaitableOperation(opName, this.client, "continue", this.sftp, "error", (complete) => {
                 if (!this.sftp) {
                     return false;
                 }
-                return !this.sftp.readdir(directory, (err, list) => {
+                let should_wait = !this.sftp.readdir(directory, (err, list) => {
                     if (err) {
                         this.logFn(LogType.debug, () => localize("debug.operation.error", "{0} error: {1}", opName, err.message));
                     } else {
                         files = list;
                     }
                     complete.release();
+                    if (should_wait) {
+                        this.client?.emit("continue");
+                    }
                 });
+                return should_wait;
             }, this.logFn);
         }
         this.waitOperation.release();
@@ -180,6 +196,7 @@ export class SftpClient extends SshClient {
                     filename: file.filename,
                     isDirectory,
                     isFile: !isDirectory,
+                    vers: 0,
                 };
                 return entry;
             });
@@ -196,20 +213,24 @@ export class SftpClient extends SshClient {
                 this.waitOperation.release();   // let getStat work
                 const stat = await this.getStat(directory);
                 await this.waitOperation.acquire();
-                if (!stat && this.enabled) {
+                if (!stat && this.client && this.enabled) {
                     const opName = localize("operation.createdir", "create directory {0} via sftp{1}", directory, this.tag ? " " + this.tag : "");
-                    await WaitableOperation(opName, this.sftp, "continue", this.sftp, "error", (complete) => {
+                    await WaitableOperation(opName, this.client, "continue", this.sftp, "error", (complete) => {
                         if (!this.sftp) {
                             return false;
                         }
-                        return !this.sftp.mkdir(directory, (err) => {
+                        let should_wait = !this.sftp.mkdir(directory, (err) => {
                             if (err) {
                                 this.logFn(LogType.debug, () => localize("debug.operation.error", "{0} error: {1}", opName, err.message));
                             } else {
                                 retCode = true;
                             }
                             complete.release();
+                            if (should_wait) {
+                                this.client?.emit("continue");
+                            }
                         });
+                        return should_wait;
                     }, this.logFn);
                 }
             }
@@ -251,7 +272,7 @@ export class SftpClient extends SshClient {
                 if (!this.client) {
                     return false;
                 }
-                return !this.client.sftp((err, sftpGot) => {
+                let should_wait = !this.client.sftp((err, sftpGot) => {
                     if (err) {
                         this.logFn(LogType.debug, () => localize("debug.operation.error", "{0} error: {1}", opName, err.message));
                     } else {
@@ -271,7 +292,11 @@ export class SftpClient extends SshClient {
                         });
                     }
                     complete.release();
+                    if (should_wait) {
+                        this.client?.emit("continue");
+                    }
                 });
+                return should_wait;
             }, this.logFn);
         }
         return this.sftp !== undefined;
